requirements.txt
scipy~=1.8.1
numpy~=1.22.3
matplotlib~=3.2.2
opencv_python==4.1.2.30
torch~=1.8.0
torchvision~=0.9.0
tqdm~=4.64.0
Pillow~=6.2.1
h5py==2.10.0

opencv-python~=4.5.5.64
pycocotools~=2.0.4
thop~=0.0.31.post2005241907
torchsummary~=1.5.1